Add files via upload
This is the file the current Office Script is based on. You'll need to create a new script and copy and paste the script. Then create and grab your API key from OpenAI and add it to Excel. You can connect your script to a button and you're all set.